fixed the insert fx - async calls were not being resolved properly because two async calls happen at once; put the 2nd async inside the .then of the first to execute in succession.
